+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0
 
-- Rename class `Carol` -> `Pattern`
+- Rename: class `Carol` -> `Pattern`
+- Rename: function `regex()` -> `pattern()`
 - Type Declaration: Add documents.
 
 ## 0.1.0